Split todo lifecycle smoke fixtures
